+++ src/main/scala/characters/Resources.scala
@@ -1,7 +1,7 @@
-package person
+package characters
 
-class GuessWhoGame {
-  //val firstCharacters: List[Person] = List(person1(name="David"))
+class Resources {
+  //create a list of characters
   val firstCharacter: Person = Person(name = "David", gender = "Male", hairColour = "Black", hairLength = "Short")
   val secondCharacter: Person = Person(name = "Sam", gender = "Male", hairColour = "Grey", hairLength = "Short")
   val thirdCharacter: Person = Person(name = "Victoria", gender = "Female", hairColour = "Purple", hairLength = "Long")
@@ -9,6 +9,9 @@ class GuessWhoGame {
 
   val charactersList: List[Person] = List(firstCharacter, secondCharacter, thirdCharacter, fourthCharacter)
 
+  def returnMapOfQuestions(): List[String] ={
+    val listOfQuestions:List[String] = List("")
+  }
   def returnListOfCharacters(): List[Person] ={
     charactersList
   }
